Revise connected() function to accept people-lists
Previously, connected() only worked on rows and columns. For clues that
specify a sequence of connected roles with respect to a person, this
was insufficient.

To be able to reason about connected sequences, the input needs to be
ordered. As using the '&' intersection operator was convenient, we
introduce a class that behaves as an ordered list, but also implements
the intersection operator.